Refactoring directory structure.
Add storage module and move cache_storage and so on under it.
Also split tests to match it.
